+++ alphashape/__init__.py
@@ -4,6 +4,10 @@ __author__ = """Kenneth E. Bellock"""
 __email__ = 'ken@bellock.net'
 
 from .alphashape import alphashape
+from .alphashape import circumradius
+from .alphashape import circumcenter
+from .alphashape import alphasimplices
 from .optimizealpha import optimizealpha
 from ._version import __version__  # noqa: F401
-__all__ = ['alphashape', 'optimizealpha']
+__all__ = ['alphashape', 'optimizealpha', 'circumradius',
+           'circumcenter', 'alphasimplices']