+++ alphashape/cli.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 """Console script for alphashape."""
+import os
 import sys
 import click
 import click_log
@@ -26,9 +27,9 @@ def main(source, target, alpha, epsg):
     """
     Example console appication using the alphashape toolbox.
 
-    Given an input shapefile INPUT with point geometry, write out a new
-    shapefile OUTPUT that contains the geometries resulting from execting the
-    alpha shape toolbox.
+    Given an input shapefile or GeoJSON INPUT with point geometry, write out a
+    new OUTPUT that contains the geometries resulting from execting the alpha
+    shape toolbox.
 
     The alpha parameter is optional.  If provided it will return the alpha
     shape for the given value, if one is not provided, the tightest fitting
@@ -38,23 +39,26 @@ def main(source, target, alpha, epsg):
     shape analysis in.  If one is not given the coordinate system of the input
     data will be used.
 
-    The output shapefile will always have the same coordinate system as the
+    The output file will always have the same coordinate system as the
     source file.
+
+    The output file format will be determined by the extension of the provided
+    target filename and can be written out in shapefile format or GeoJSON.
     """
     # Read in source data
     source_filename = click.format_filename(source)
     target_filename = click.format_filename(target)
-    LOGGER.info('Reading source shapefile: %s', source_filename)
+    LOGGER.info('Reading source file: %s', source_filename)
     try:
         gdf = geopandas.read_file(source_filename)
     except:  # noqa: E722
-        LOGGER.error('Could not read source shapefile')
+        LOGGER.error('Could not read source file')
         return 10
 
     # Source data type checking
     if not any([isinstance(
             p, shapely.geometry.Point) for p in gdf['geometry']]):
-        LOGGER.error('Source shapefile does not contain multipiont features')
+        LOGGER.error('Source file does not contain multipiont features')
         return 20
 
     # Project data if given an EPSG code
@@ -63,7 +67,7 @@ def main(source, target, alpha, epsg):
         try:
             gdf_input = gdf.to_crs({'init': 'epsg:%s' % epsg})
         except:  # noqa: E722
-            LOGGER.error('Could not project source shapefile')
+            LOGGER.error('Could not project source data')
             return 30
     else:
         gdf_input = gdf
@@ -85,12 +89,15 @@ def main(source, target, alpha, epsg):
             LOGGER.error('Could not project alpha shape')
             return 50
 
-    # Write out the target shapefile
-    LOGGER.info('Writing target shapefile: %s', target_filename)
+    # Write out the target file
+    LOGGER.info('Writing target file: %s', target_filename)
     try:
-        alpha_shape.to_file(target)
+        if os.path.splitext(target)[1].lower() == '.geojson':
+            alpha_shape.to_file(target, driver='GeoJSON')
+        else:
+            alpha_shape.to_file(target)
     except:  # noqa: E722
-        LOGGER.error('Could not write target shapefile')
+        LOGGER.error('Could not write target file')
         return 60
     return 0